Feature/Favorites

+++ alembic/versions/da44ff3b0455_add_overview_image_to_property.py
@@ -9,6 +9,7 @@ Create Date: 2026-01-20 22:44:33.631957
 from typing import Sequence, Union
 
 from alembic import op
+from sqlalchemy import inspect
 import sqlalchemy as sa
 
 
@@ -20,10 +21,19 @@ depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
-    op.add_column(
-        "properties", sa.Column("overview_image", sa.String(length=500), nullable=True)
-    )
+    conn = op.get_bind()
+    inspector = inspect(conn)
+    columns = [c["name"] for c in inspector.get_columns("properties")]
+    if "overview_image" not in columns:
+        op.add_column(
+            "properties",
+            sa.Column("overview_image", sa.String(length=500), nullable=True),
+        )
 
 
 def downgrade() -> None:
-    op.drop_column("properties", "overview_image")
+    conn = op.get_bind()
+    inspector = inspect(conn)
+    columns = [c["name"] for c in inspector.get_columns("properties")]
+    if "overview_image" in columns:
+        op.drop_column("properties", "overview_image")
